experiments/requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.7
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.13.1
gitdb==4.0.11
GitPython==3.1.40
hydra-core==1.3.2
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
lit==17.0.4
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
packaging==23.2
pandas==2.0.3
pathtools==0.1.2
Pillow==10.1.0
protobuf==4.25.0
psutil==5.9.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sentry-sdk==1.34.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wandb==0.15.12
zipp==3.17.0
